Move css! to silkenweb-style-macros
