+++ Cargo.toml
@@ -27,6 +27,7 @@ version = "0.5.0"
 # main dependencies
 silkenweb = { path = "packages/silkenweb", version = "0.5.0" }
 silkenweb-style = { path = "packages/style", version = "0.5.0" }
+silkenweb-style-macros = { path = "packages/style-macros", version = "0.5.0" }
 silkenweb-signals-ext = { path = "packages/signals-ext", version = "0.5.0" }
 silkenweb-macros = { path = "packages/macros", version = "0.5.0" }
 silkenweb-tauri-proc-macro = { path = "packages/tauri-proc-macro", version = "0.5.0" }
