+++ packages/style-macros/Cargo.toml
@@ -1,22 +1,26 @@
 [package]
-name = "silkenweb-macros"
+name = "silkenweb-style-macros"
 version = { workspace = true }
 authors = ["Simon Bourne <simonbourne@gmail.com>"]
 edition = "2021"
-description = "Silkenweb proc macros"
+description = "Macros for silkenweb-style"
 homepage = "https://github.com/silkenweb/silkenweb"
 repository = "https://github.com/silkenweb/silkenweb"
 license = "MIT OR Apache-2.0"
 
-categories = ["gui"]
-keywords = ["css"]
+categories = ["gui", "web-programming"]
+keywords = ["reactive", "web", "html", "browser", "dom"]
 
 [lib]
 proc-macro = true
 
 [dependencies]
-syn = { workspace = true, features = ["full"] }
+silkenweb-style = { workspace = true }
+cssparser = { workspace = true }
+lightningcss = { workspace = true }
 proc-macro-error = { workspace = true }
-quote = { workspace = true }
 proc-macro2 = { workspace = true }
+syn = { workspace = true, features = ["full"] }
+quote = { workspace = true }
+grass = { workspace = true }
 derive_more = { workspace = true }
